requirements.txt
asgiref==3.5.2
backcall==0.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
ChatterBot==1.0.8
chatterbot-corpus==1.2.0
click==8.1.3
colorama==0.4.6
confection==0.0.3
cryptography==38.0.4
cymem==2.0.7
decorator==5.1.1
Deprecated==1.2.13
Django==3.2.16
django-bootstrap-form==3.4
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.1.0/en_core_web_md-2.1.0.tar.gz
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
gcloud==0.18.3
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.1.0
ipython==7.34.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==2.6.0
jwcrypto==1.4.2
langcodes==3.3.0
MarkupSafe==2.1.1
mathparse==0.1.2
matplotlib-inline==0.1.6
murmurhash==1.0.9
numpy==1.21.6
oauth2client==4.1.3
packaging==21.3
parso==0.8.3
pathy==0.10.0
pickleshare==0.7.5
Pillow==9.3.0
plac==0.9.6
preshed==3.0.8
prompt-toolkit==3.0.33
protobuf==4.21.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
Pyrebase4==4.6.0
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2022.6
PyYAML==3.13
requests==2.28.1
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
smart-open==5.2.1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.3.24
sqlparse==0.4.3
srsly==2.4.5
thinc==8.1.5
tqdm==4.64.1
traitlets==5.5.0
typer==0.7.0
typing_extensions==4.1.1
urllib3==1.26.13
wasabi==0.10.1
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.10.0